WIP better openurl handling
This is a work-in-progress to improve reliability on the handling of
opening urls.

So far I was trying to make the interface more unified and cleaner to
help make the outcomes more manageable, predictable, and force a
well-defined outcome at compile-time (even if that means showing a loading screen or an
error screen).

This is being done to avoid silent failures and error paths that are
hard to see within convoluted logic paths.