Improve efficiency of event holder insertion
This commit improves the efficiency of the algorithm used for sorted
insertion on EventHolder.

Previously, it used an O(n) algorithm, which appears to be slow on large
accounts, and blocks the main thread — occasionally causing the app to
freeze and be terminated by the watchdog.

The events array was also changed to be only settable by the EventHolder
class, to prevent breaking the assumption that the list needs to be
pre-sorted prior to insertion.

Changelog-Changed: Improved efficiency of the app to reduce freezes on large accounts
Signed-off-by: Daniel D’Aquino <daniel@daquino.me>